+++ lib/zhihu_api_parsers.dart
@@ -7,4 +7,5 @@ library;
 
 export 'src/comment_emoticons.dart';
 export 'src/json_tools.dart';
+export 'src/search_hot.dart';
 export 'src/search_suggestions.dart';